Split Terraform custom code to multiple files
